refactor and move things around
