+++ twig_server/models/db_objects.py
@@ -44,6 +44,19 @@ class BaseDbObject(Generic[VK], BaseModel):
         }
 
     def serialize_for_neo4j_insert(self):
+        """
+        Serializes this object for insertion into the database.
+        In particular, this removes the ID field, since that is
+        automatically generated by the database.
+
+        Returns:
+            A dictionary that can be inserted into the database.
+
+        Notes:
+            This is a bit hacky, since pydantic doesn't have a way to just serialize
+            datetime objects when you call dict. If performance starts being
+            a concern, replace this method with a manual implementation.
+        """
         return json.loads(self.json(exclude={"id"}))
