+++ twig_server/neo4j_orm_lite/__init__.py
@@ -1,0 +1,10 @@
+"""
+My attempt at building a lightweight and simple ORM for Neo4j.
+
+Goals:
+    - Be generic enough s.t. we can extract this into a separate package and repo.
+    - Powerful API that is simple and just works - we want a better UI than Apple.
+    - Performant
+"""
+
+__version__ = "0.0.0-dev"
