+++ twig_server/neo4j_orm_lite/orm.py
@@ -1,77 +1,10 @@
-import enum
-from typing import Optional, ClassVar, Generic
+from typing import ClassVar, Generic
 
 import neo4j
 
 from twig_server.models.db_objects import BaseDbObject
 from twig_server.models.types import VK
-
-
-class Operation(enum.Enum):
-    CREATE = "CREATE"
-    READ = "READ"
-    UPDATE = "UPDATE"
-    DELETE = "DELETE"
-
-
-class CypherQueryString:
-    def __init__(
-            self,
-            operation: Operation,
-            *,
-            returned_idx_name: Optional[str] = None,
-            is_single: Optional[bool] = None,
-            required_parameters: Optional[list[str]] = None,
-            query_string: str,
-    ):
-        # TODO: Checks
-        self.operation = operation
-        self.required_parameters = required_parameters
-        self.query_string = query_string
-        self.is_single = is_single
-        self.returned_idx_name = returned_idx_name
-
-    def __str__(self):
-        return self.query_string
-
-    def execute(self, sess: neo4j.Session, parameters: Optional[dict] = None):
-        # performs validation as well
-        # Return the single thing if self.is_single otherwise it will return the result
-
-        if self.required_parameters:
-            if len(self.required_parameters) > 0 and parameters is None:
-                raise ValueError("Parameters are required for this query.")
-            if set(self.required_parameters) != set(parameters.keys()):
-                raise ValueError(
-                    f"Parameters {parameters.keys()} do not match"
-                    f" required parameters {self.required_parameters}"
-                )
-
-        result = sess.run(self.query_string, parameters)
-        if self.is_single:
-            result = result.single()
-            if not result:
-                raise Exception("A more concrete exception would be nice")
-            return result[self.returned_idx_name]
-        else:
-            return result
-
-
-class CypherQuery:
-    def __init__(self, query_string: CypherQueryString, obj: BaseDbObject):
-        self.query_string = query_string
-        self.obj = obj
-
-    def execute(self, sess: neo4j.Session, parameters: Optional[dict] = None) -> M:
-        result = self.query_string.execute(sess, parameters)
-        if self.query_string.is_single:
-            return self.desired_type(id=result.id, **dict(result.items()))
-        else:
-            ret_val = []
-            for record in result:
-                n = record[self.query_string.returned_idx_name]
-                ret_val.append(self.desired_type(id=n.id, **dict(n.items())))
-            return ret_val
+from twig_server.neo4j_orm_lite.query import CypherQueryString, Operation
 
 
 class TwigNeoModel(Generic[VK]):
@@ -175,6 +108,7 @@ class TwigORMSession:
         >>>    sess.add(proj)
         >>>    sess.commit()
     """
+
     def __init__(self, driver: neo4j.Neo4jDriver):
         self.driver = driver
